invfilter

+++ src/combat.cc
@@ -3232,7 +3232,7 @@ static int _combat_turn(Object* obj, bool a2)
 {
     _combat_turn_obj = obj;
 
-    RunCombatTurnHook((long)obj, a2);
+    RunCombatTurnHook(obj, a2);
 
     attackInit(&_main_ctd, obj, nullptr, HIT_MODE_PUNCH, HIT_LOCATION_TORSO);
